Refactor hexagonal  | ThreadPoolProvider
