+++ .env.example
@@ -9,6 +9,7 @@ LOG_FORMAT='[%(asctime)s.%(msecs)03d] - %(levelname)s - %(message)s'
 APP_NAME='Jabba AI-Bot'
 APP_VERSION=0.0.0.dev
 APP_API_CORS_ALLOWED_DOMAINS=*
+APP_MAX_THREAD_POOL_WORKERS=20
 
 # ----------------------------
 #   Bot
